requirements.txt
# Django Core 
Django>=5.0,<5.1 
python-decouple>=3.8 
psycopg2-binary>=2.9
dj-database-url>=2.1.0 
django-allauth>=0.57 
dj-stripe>=2.8 
django-paypal>=2.0 
djangorestframework>=3.14 
django-cors-headers>=4.3 
Pillow>=10.0 
pandas>=2.2 
numpy>=1.26 
plotly>=5.22

# Production Server
gunicorn>=21.2.0
whitenoise>=6.6.0 
